+++ Application/Features/Products/Queries/GetAllProducts/GetAllProductsFilter.cs
@@ -1,0 +1,12 @@
+﻿using Application.Filters;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Application.Features.Products.Queries.GetAllProducts
+{
+    public class GetAllProductsFilter : PaginationFilter
+    {
+
+    }
+}