+++ Application/Filters/PaginationFilter.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Application.Filters
+{
+    public class PaginationFilter
+    {
+        public int PageNumber { get; set; }
+        public int PageSize { get; set; }
+        public PaginationFilter()
+        {
+            this.PageNumber = 1;
+            this.PageSize = 10;
+        }
+        public PaginationFilter(int pageNumber, int pageSize)
+        {
+            this.PageNumber = pageNumber < 1 ? 1 : pageNumber;
+            this.PageSize = pageSize > 10 ? 10 : pageSize;
+        }
+    }
+}