+++ Application/Interfaces/IGenericRepositoryAsync.cs
@@ -8,7 +8,7 @@ namespace Application.Interfaces
     public interface IGenericRepositoryAsync<T> where T : class
     {
         Task<T> GetByIdAsync(int id);
-        Task<IReadOnlyList<T>> ListAllAsync();
+        Task<IReadOnlyList<T>> GetAllAsync(int pageNumber, int pageSize);
         Task<T> AddAsync(T entity);
         Task UpdateAsync(T entity);
         Task DeleteAsync(T entity);