+++ Application/Mappings/GeneralProfile.cs
@@ -14,6 +14,7 @@ namespace Application.Mappings
         {
             CreateMap<Product, GetAllProductsViewModel>().ReverseMap();
             CreateMap<CreateProductCommand, Product>();
+            CreateMap<GetAllProductsQuery, GetAllProductsFilter>();
         }
     }
 }