+++ Application/Wrappers/PagedResponse.cs
@@ -9,9 +9,6 @@ namespace Application.Wrappers
         public int PageNumber { get; set; }
         public int PageSize { get; set; }
 
-        public int TotalPages { get; set; }
-        public int TotalRecords { get; set; }
-
         public PagedResponse(T data, int pageNumber, int pageSize)
         {
             this.PageNumber = pageNumber;