+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - custom schema and names for Identity Table
 - added automapper with general profile
 - added endpoints and handlers for product /getall and /post
+- added pagination filters and implemented pagination for GetAll Method of Generic Repository
 - and more.
 
 ### Changed